+++ CHANGES.rst
@@ -6,6 +6,8 @@ Changelog
 
 - added `logging_level` parameter to the `start` method of the `weka.core.jvm` module, enabling the user
   to turn off debugging output in an easy way (https://github.com/fracpete/python-weka-wrapper3/issues/40)
+- added method `cv_splits` to class `Instances` from module `weka.core.dataset` to return a list of
+  train/test tuples as used by cross-validation
 - ...